rename re to rec
're' is also the name of a built-in Python module.
